Check for duplicate message earlier
Check if we have seen the notification message before any processing
happens on it

+++ dot/types/block_data.go
@@ -28,7 +28,7 @@ func NewEmptyBlockData() *BlockData {
 
 // Number returns the block header number.
 func (bd *BlockData) Number() uint {
-	return bd.Header.Number
+	return uint(bd.Header.Number)
 }
 
 func (bd *BlockData) String() string {
